# swagger.yml
swagger: '2.0'
info:
  description: This is an API for accessing all sorts of college football data.  Please note that API keys should be supplied with "Bearer " prepended (e.g. "Bearer your_key"). API keys can be acquired from the CollegeFootballData.com website.
  version: 4.3.2
  title: College Football Data API
  contact:
    email: admin@collegefootballdata.com
host: api.collegefootballdata.com
basePath: /
...
schemes:
  - https
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization
security:
  - ApiKeyAuth: []
paths:
...
  /conferences:
    get:
      tags:
        - conferences
      summary: Conferences
      description: Get conference list
      operationId: getConferences
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Conference'
        '400':
          description: error
  /teams:
    get:
      tags:
        - teams
      summary: Team information
      description: Get team information
      operationId: getTeams
      produces:
        - application/json
      parameters:
        - name: conference
          description: Conference abbreviation filter
          in: query
          required: false
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Team'
        '400':
          description: error
  /teams/fbs:
    get:
      tags:
        - teams
      summary: FBS team list
      description: Information on major division teams
      operationId: getFbsTeams
      produces:
        - application/json
      parameters:
        - name: year
          description: Year filter
          in: query
          required: false
          type: integer
          minimum: 1869
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Team'
        '400':
          description: error
  /roster:
    get:
      tags:
        - teams
      summary: Team rosters
      description: Roster data
      operationId: getRoster
      produces:
        - application/json
      parameters:
        - name: team
          description: Team name
          in: query
          required: false
          type: string
        - name: year
          description: Season year
          in: query
          required: false
          type: integer
          minumum: 2009
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Player'
        '400':
          description: error
...
definitions:
...
  Conference:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      short_name:
        type: string
      abbreviation:
        type: string
      classification:
        type: string
  Team:
    type: object
    properties:
      id:
        type: integer
      school:
        type: string
      mascot:
        type: string
      abbreviation:
        type: string
      alt_name_1:
        type: string
      alt_name_2:
        type: string
      alt_name_3:
        type: string
      conference:
        type: string
      division:
        type: string
      color:
        type: string
      alt_color:
        type: string
      logos:
        type: array
        items:
          type: string
      location:
        type: object
        properties:
          venue_id:
            type: integer
          name:
            type: string
          city:
            type: string
          state:
            type: string
          zip:
            type: string
          country_code:
            type: string
          timezone:
            type: string
          latitude:
            type: number
          longitude:
            type: number
          elevation:
            type: number
          capacity:
            type: number
          year_constructed:
            type: number
          grass:
            type: boolean
          dome:
            type: boolean
  Player:
    type: object
    properties:
      id:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      team:
        type: string
      height:
        type: integer
      weight:
        type: integer
      jersey:
        type: integer
      year:
        type: integer
      position:
        type: string
      home_city:
        type: string
      home_state:
        type: string
      home_country:
        type: string
      home_latitude:
        type: number
      home_longitude:
        type: number
      home_county_fips:
        type: string
      recruit_ids:
        type: array
        items:
          type: integer
...